project5-SZU-python/assignments/XImageSearch/requirements.txt
django==4.2.11
numpy==1.26.4
scipy==1.11.4
pillow==10.4.0
anthropic>=0.20.0
